Refactored numberFormat util to plain funcs
