+++ src/components/FuelSavingsResults.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import NumberFormatter from '../utils/numberFormatter';
+import {scrubFormatting} from '../utils/numberFormat';
 
 // This is a stateless functional component. (Also known as pure or dumb component)
 // More info: https://facebook.github.io/react/blog/2015/10/07/react-v0.14.html#stateless-functional-components
 // And https://medium.com/@joshblack/stateless-components-in-react-0-14-f9798f8b992d
 // Props are being destructured below to extract the savings object to shorten calls within component.
 const FuelSavingsResults = ({savings}) => {
-  const savingsExist = NumberFormatter.scrubFormatting(savings.monthly) > 0;
+  const savingsExist = scrubFormatting(savings.monthly) > 0;
   const savingsClass = savingsExist ? 'savings' : 'loss';
   const resultLabel = savingsExist ? 'Savings' : 'Loss';
